requirements.txt
streamlit
opencv-python-headless
mediapipe==0.10.14
numpy
Pillow